+++ board/board-application/src/main/java/com/seatliberator/seatliberator/board/infrastructure/security/BoardCapability.java
@@ -1,6 +1,6 @@
 package com.seatliberator.seatliberator.board.infrastructure.security;
 
-import com.seatliberator.seatliberator.identity.client.role.Capability;
+import com.seatliberator.seatliberator.identity.core.role.Capability;
 import lombok.RequiredArgsConstructor;
 
 @RequiredArgsConstructor
