+++ bootstrap/resource-application-starter/src/main/java/com/seatliberator/seatliberator/bootstrap/security/ActorContextJwtAuthenticationConverter.java
@@ -1,9 +1,9 @@
 package com.seatliberator.seatliberator.bootstrap.security;
 
-import com.seatliberator.seatliberator.identity.client.role.Capability;
 import com.seatliberator.seatliberator.identity.client.role.NamespaceRoleCapabilitiesRegistry;
 import com.seatliberator.seatliberator.identity.core.actor.Actor;
 import com.seatliberator.seatliberator.identity.core.actor.SimpleActor;
+import com.seatliberator.seatliberator.identity.core.role.Capability;
 import com.seatliberator.seatliberator.identity.core.role.NamespaceRole;
 import com.seatliberator.seatliberator.identity.core.role.NamespaceRoleDeserializer;
 import com.seatliberator.seatliberator.identity.core.role.Role;
