+++ identity/identity-client/build.gradle.kts
@@ -21,6 +21,7 @@ dependencies {
     implementation("org.slf4j:slf4j-api")
 
     // Test
+    testImplementation(project(":identity:identity-core"))
     testImplementation("org.springframework.boot:spring-boot-starter-test")
     testImplementation(platform("org.junit:junit-bom:5.10.0"))
     testImplementation("org.junit.jupiter:junit-jupiter")
