--- identity/identity-client/src/main/java/com/seatliberator/seatliberator/identity/client/role/RoleCapabilities.java
@@ -1,11 +1,0 @@
-package com.seatliberator.seatliberator.identity.client.role;
-
-import com.seatliberator.seatliberator.identity.core.role.Role;
-
-import java.util.Set;
-
-public record RoleCapabilities(
-        Role role,
-        Set<Capability> capabilities
-) {
-}
