+++ identity/identity-client/src/test/java/com/seatliberator/seatliberator/identity/client/role/NamespaceRoleCapabilitiesRegistryTest.java
@@ -1,6 +1,8 @@
 package com.seatliberator.seatliberator.identity.client.role;
 
 import com.seatliberator.seatliberator.identity.core.role.Role;
+import com.seatliberator.seatliberator.identity.core.role.RoleCapabilities;
+import com.seatliberator.seatliberator.identity.core.role.SimpleCapability;
 import com.seatliberator.seatliberator.identity.core.role.SimpleNamespaceRole;
 import com.seatliberator.seatliberator.kernel.SimpleApplicationNamespace;
 import org.junit.jupiter.api.DisplayName;
