+++ identity/identity-core/src/main/java/com/seatliberator/seatliberator/identity/core/role/Capability.java
@@ -1,4 +1,4 @@
-package com.seatliberator.seatliberator.identity.client.role;
+package com.seatliberator.seatliberator.identity.core.role;
 
 public interface Capability {
     String scope();
