+++ identity/identity-core/src/main/java/com/seatliberator/seatliberator/identity/core/role/RoleCapabilities.java
@@ -1,0 +1,9 @@
+package com.seatliberator.seatliberator.identity.core.role;
+
+import java.util.Set;
+
+public record RoleCapabilities(
+        Role role,
+        Set<Capability> capabilities
+) {
+}
