+++ identity/identity-core/src/main/java/com/seatliberator/seatliberator/identity/core/role/SimpleCapability.java
@@ -1,4 +1,4 @@
-package com.seatliberator.seatliberator.identity.client.role;
+package com.seatliberator.seatliberator.identity.core.role;
 
 public record SimpleCapability(String scope, String description) implements Capability {
 }
