+++ reservation/reservation-web-mvc/src/main/java/com/seatliberator/seatliberator/reservation/infrastructure/web/shared/security/ReservationRoleCapabilityConfiguration.java
@@ -1,7 +1,7 @@
 package com.seatliberator.seatliberator.reservation.infrastructure.web.shared.security;
 
-import com.seatliberator.seatliberator.identity.client.role.RoleCapabilities;
 import com.seatliberator.seatliberator.identity.core.role.Role;
+import com.seatliberator.seatliberator.identity.core.role.RoleCapabilities;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.security.config.annotation.method.configuration.EnableMethodSecurity;
